Kjør tester med redis testcontainer
Appen er avhengig av healthchecks og koin crasher når redis ikke klarer
å koble til.
